refactor: fold the docstring reader into the hvPlot views module
Reading hvPlot's option descriptions is only useful to the params
declared next to it, so it does not need a module of its own.
